+++ lib/flutter_split_navigator.dart
@@ -1,0 +1,3 @@
+export 'src/split_navigator.dart';
+export 'src/split_observers.dart';
+export 'src/split_handler.dart';