+++ lib/flutter_split_view.dart
@@ -1,3 +1,0 @@
-export 'src/split_view.dart';
-export 'src/observers.dart';
-export 'src/split_handler.dart';