+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 ## 1.0.0
 
-* Implemented `FlutterSplitView` widget for creating a split view layout.
+* Implemented `FlutterSplitNavigator` widget for creating a split view layout.
 * Added responsive layout that adapts to screen width.
 * Implemented primary and secondary navigation.
 * Added placeholder widget for the secondary view.
-* Implemented `_RouteObserver` to trigger rebuilds on route changes.
 
 ## 0.0.1